refactor: refactor user module structure
